requirements_prod.txt
# === Core Dependencies ===
setuptools
wheel
python-dotenv
pydantic

# === Google Cloud ===
google-api-core==2.8.2
google-auth-oauthlib==1.2.1
google-cloud-bigquery
google-cloud-storage==2.14.0
googleapis-common-protos==1.56.4
gcsfs==2025.3.0
db-dtypes
pyarrow

# === Numeric/ML Foundation ===
numpy==1.26.4
scipy==1.15.2
protobuf==3.19.6
h5py==3.10.0

# === Data Handling ===
pandas==1.5.3

# === Machine Learning ===
scikit-learn==1.6.1
scikit-survival==0.24.0
shap==0.47.1

# === ML Operations ===
mlflow==2.3.2
prefect==2.19.2

# === API Framework ===
fastapi==0.108.0
uvicorn
aiohttp==3.8.6
aiohappyeyeballs==2.6.1
aiosignal==1.3.2
frozenlist==1.5.0
multidict==6.2.0
yarl==1.18.3
httpx<0.28
pytz
Requests
